pyproject.toml
[tool.poetry]
name = "back"
version = "0.1.0"
description = "Backend api for blog"
authors = ["Haruki Ambai <haruru8998@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.89.1"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
psycopg2-binary = "^2.9.5"
sqlalchemy = "^2.0.0"
psycopg2 = "^2.9.9"
boto3 = "^1.28.68"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
